sdk/sqreen-instrumentation-tool: move validation tests into the sqhook package
