+++ agent/sqlib/sqhook/test/testdata/hello-example/main.go
@@ -7,7 +7,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/sqreen/go-agent/sdk/instrumentation/sqreen/test/testdata/helpers"
+	"github.com/sqreen/go-agent/agent/sqlib/sqhook/test/testdata/helpers"
 )
 
 // no params nor results
