+++ agent/sqlib/sqhook/test/testdata/hello-world/main.go
@@ -7,7 +7,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/sqreen/go-agent/sdk/instrumentation/sqreen/test/testdata/helpers"
+	"github.com/sqreen/go-agent/agent/sqlib/sqhook/test/testdata/helpers"
 )
 
 func main() {
